refactor(capture): rejoin the buffer, accessor, websocket and network clusters
Four merges, each backed by a graph edge the filenames obscured:

  buffer_store.go <- memory.go
    memory.go is filed as a Capture concern but half of it is BufferStore's own eviction
    accounting: buffer_store.go and memory.go referenced each other (wsEventMemory /
    nbEntryMemory / calcWSMemory), so every append already crossed the file boundary.
    The two public Capture memory accessors come along; they are three lines each.

  accessors.go <- accessor_counts_health.go, accessor_events.go, accessor_performance.go
    Commit 0e8e2cd5 created all four accessor_* files at once. They are one read layer:
    take the read lock, delegate to a store, return. accessor.go (the fifth) was a
    comment-only stub describing that split and went in the previous commit.

  websocket.go <- websocket_handlers.go, websocket_status.go
    websocket_status.go is down to the single GetWebSocketStatus accessor now that the
    connection tracker lives in capture/wsconn.

  network.go <- network_waterfall.go, network_waterfall_store.go
    network_bodies.go and network_waterfall*.go are the two halves of one ingest path
    and are the only readers of NetworkWaterfallBuffer.

Largest resulting file is 342 lines. No declaration changed.

internal/capture: 44 -> 37 non-test files.
