--- internal/capture/memory.go
@@ -1,60 +1,0 @@
-// Purpose: Implements capture buffer memory accounting helpers and O(1) memory-total accessors.
-// Why: Keeps ingestion and eviction logic memory-aware to prevent runaway daemon usage.
-// Docs: docs/features/feature/backend-log-streaming/index.md
-
-package capture
-
-const (
-	// Per-entry memory estimates
-	wsEventOverhead     = 200 // bytes overhead per WS event
-	networkBodyOverhead = 300 // bytes overhead per network body
-	actionMemoryFixed   = 500 // bytes per enhanced action (fixed estimate)
-)
-
-// ============================================
-// Per-Entry Memory Calculation
-// ============================================
-
-// wsEventMemory returns the memory estimate for a single WS event.
-func wsEventMemory(e *WebSocketEvent) int64 {
-	return int64(len(e.Data)) + wsEventOverhead
-}
-
-// nbEntryMemory returns the memory estimate for a single network body entry.
-func nbEntryMemory(b *NetworkBody) int64 {
-	return int64(len(b.RequestBody)+len(b.ResponseBody)) + networkBodyOverhead
-}
-
-// ============================================
-// Per-Buffer Memory Accessors (caller must hold lock)
-// ============================================
-
-// calcWSMemory returns the running total of WS buffer memory (caller must hold lock).
-// O(1) — maintained incrementally by add/evict/clear operations.
-func (s *BufferStore) calcWSMemory() int64 {
-	return s.wsMemoryTotal
-}
-
-// calcNBMemory returns the running total of network bodies buffer memory (caller must hold lock).
-// O(1) — maintained incrementally by add/evict/clear operations.
-func (s *BufferStore) calcNBMemory() int64 {
-	return s.networkBodyMemoryTotal
-}
-
-// ============================================
-// Public Memory Accessors
-// ============================================
-
-// GetWebSocketBufferMemory returns approximate memory usage of WS buffer
-func (c *Capture) GetWebSocketBufferMemory() int64 {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.buffers.calcWSMemory()
-}
-
-// GetNetworkBodiesBufferMemory returns approximate memory usage of network bodies buffer
-func (c *Capture) GetNetworkBodiesBufferMemory() int64 {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.buffers.calcNBMemory()
-}
