+++ internal/capture/websocket.go
@@ -1,10 +1,12 @@
-// Purpose: Implements websocket event ingestion, repair, filtering, and query handlers for capture buffers.
+// Purpose: Implements websocket event ingestion, filtering, HTTP handlers, and the connection-status accessor.
 // Why: Preserves websocket lifecycle/message evidence with consistent buffering and binary-format enrichment.
 // Docs: docs/features/feature/backend-log-streaming/index.md
 
 package capture
 
 import (
+	"encoding/json"
+	"net/http"
 	"strings"
 	"time"
 
@@ -113,4 +115,41 @@ func (c *Capture) GetWebSocketEvents(filter WebSocketEventFilter) []WebSocketEve
 		}
 	}
 	return filtered
+}
+
+// HandleWebSocketEvents handles POST /websocket-events from the extension.
+// Reads go through GET /telemetry?type=websocket_events.
+func (c *Capture) HandleWebSocketEvents(w http.ResponseWriter, r *http.Request) {
+	if !util.RequireMethod(w, r, "POST") {
+		return
+	}
+	body, ok := c.readIngestBody(w, r)
+	if !ok {
+		return
+	}
+	var payload struct {
+		Events []WebSocketEvent `json:"events"`
+	}
+	if err := json.Unmarshal(body, &payload); err != nil {
+		util.JSONResponse(w, http.StatusBadRequest, map[string]string{"error": "Invalid JSON"})
+		return
+	}
+	if !c.recordAndRecheck(w, len(payload.Events)) {
+		return
+	}
+	c.AddWebSocketEvents(payload.Events)
+	w.WriteHeader(http.StatusOK)
+}
+
+// HandleWebSocketStatus handles GET /websocket-status
+func (c *Capture) HandleWebSocketStatus(w http.ResponseWriter, r *http.Request) {
+	status := c.GetWebSocketStatus(WebSocketStatusFilter{})
+	util.JSONResponse(w, http.StatusOK, status)
+}
+
+// GetWebSocketStatus returns current connection states
+func (c *Capture) GetWebSocketStatus(filter WebSocketStatusFilter) WebSocketStatusResponse {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.wsConnections.Status(filter)
 }
